add a section table to the index
We were calculating section sizes based on offsets of the next section.
This broke due to not being updated for new sections... but is also
fragile as it's sensitive to writing the sections out in order.

Replace with an offset+length table.
Also write the length of the table.
Switch to using slices into the mmap region everywhere (instead of
doing offset math every time).  Bonus safety as the runtime now
does bounds checks for us even within a section.