faster repo filter
When repos are selected via the dropdown, we can do exact match filtering.
This is much faster than using a huge regex for 100s of repos.

(TODO: we could further optimize search by range limiting the found
file IDs since all files are prefixed with their repo and are sorted.)