Move indexing into server process.
This simplifies deployment and allows having push notifications
refresh either the github metadata or the repos themselves.

All of the commands are removed.
It may be useful to restore the manual-index command at a later date.

The old index mmap is now unmapped, and file closed, to stop file leaks.

Config is changed to separate the concept of repos and repo sources.