Speed up HTML rendering
Rewrite the HTML writer for syntax highlighting.
Reduce use of classes, especially nested selectors, as that's slow to
render. Instead, we inline the styles, but to support dark mode,
we only inline var references, and define the vars in the parent class
(which can be then use a media selector to limit itself to light/dark
mode).  Some classes are still used for background, line numbers, etc.