HTMX: SSE history fixes
Don't thrash history with every updated element.  
This avoids a hang in safari that shows up under heavy load.

When navigating history, correctly use the header overrides so that 
(in this case) we don't get SSE events back when doing a full reload 
(due to the backend thinking we're doing a regular HTMX request).